refractor filter in components
